+++ ubo_app/services/050-vscode/login_page.py
@@ -15,7 +15,7 @@ from ubo_gui.constants import DANGER_COLOR
 from ubo_gui.page import PageWidget
 
 from ubo_app.logging import logger
-from ubo_app.store.core import CloseApplicationAction
+from ubo_app.store.core.types import CloseApplicationAction
 from ubo_app.store.main import store
 from ubo_app.store.services.notifications import (
     Chime,